Exporter: append extra layer for nydus compression type
For nydus image, appending an extra nydus bootstrap layer
to the manifest, this layer represents the whole metadata
of filesystem view for the entire image.

Nydus uses this extra boostrap layer to reduce the number
of HTTP request and FUSE overlay's overhead on preparing
the container's rootfs.

For now we just support this in image type exporter.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>